Clean up rules and update autostart functions
Combine rules and minify them using rule_any.
run_once was not working for some processes that spawned others, which made it
impossible for pgrep to relate it back to the original process and therefore
find it. Create a new function, process_exists, to manually check if a process
exists before executing it in cases where it is necessary. Also update run_once
to use this new function.